+++ changelog.md
@@ -16,6 +16,7 @@
 * Recognise `nosuchnick` irc server message
 * Purge last remains of (long removed) mumble integration (#681, #682)
 * Don't use eval in legacy updater (#715, #716)
+* Fix crash on opening files in explorer on non-Windows (#654, #655)
 
 Contributors:
   - Wesmania